We have to use PRId64 (signed 8B integer), formats when writing the metadata records, because that's what the configuration code can handle. Configuration uses strtoll() to parse these strings, which means if we write UINT64_MAX as a PRIu64, we might get a saturated value (2^63), back.  (And of course, who knows what other strtoll() implementations do, it seems to me that an ERANGE failure isn't out of the question.)
This is a bug in existing metadata files in the field, that is, if any
of the fields we write are larger than 2^63, we'll parse them incorrectly.

Fortunately, the fields of interest are: the time (ckpt->sec), the size
(ckpt->ckpt_size), and the write-generation (ckpt->write_gen), none of
which are going to exceed 2^63 in any existing install.

+++ src/meta/meta_ckpt.c
@@ -442,19 +442,24 @@ __wt_meta_ckptlist_set(WT_SESSION_IMPL *session,
 
 		__wt_timestamp_addr_check(session, ckpt->oldest_start_ts,
 		    ckpt->newest_start_ts, ckpt->newest_stop_ts);
+
+		/*
+		 * Use PRId64 formats: WiredTiger's configuration code handles
+		 * signed 8B values.
+		 */
 		if (strcmp(ckpt->name, WT_CHECKPOINT) == 0)
 			WT_ERR(__wt_buf_catfmt(session, buf,
 			    "%s%s.%" PRId64 "=(addr=\"%.*s\",order=%" PRId64
-			    ",time=%" PRIuMAX
-			    ",size=%" PRIu64
-			    ",oldest_start_ts=%" PRIu64
-			    ",newest_start_ts=%" PRIu64
-			    ",newest_stop_ts=%" PRIu64
-			    ",write_gen=%" PRIu64 ")",
+			    ",time=%" PRId64
+			    ",size=%" PRId64
+			    ",oldest_start_ts=%" PRId64
+			    ",newest_start_ts=%" PRId64
+			    ",newest_stop_ts=%" PRId64
+			    ",write_gen=%" PRId64 ")",
 			    sep, ckpt->name, ckpt->order,
 			    (int)ckpt->addr.size, (char *)ckpt->addr.data,
 			    ckpt->order,
-			    ckpt->sec,
+			    (int64_t)ckpt->sec,
 			    ckpt->ckpt_size,
 			    ckpt->oldest_start_ts,
 			    ckpt->newest_start_ts,
@@ -463,16 +468,16 @@ __wt_meta_ckptlist_set(WT_SESSION_IMPL *session,
 		else
 			WT_ERR(__wt_buf_catfmt(session, buf,
 			    "%s%s=(addr=\"%.*s\",order=%" PRId64
-			    ",time=%" PRIuMAX
-			    ",size=%" PRIu64
-			    ",oldest_start_ts=%" PRIu64
-			    ",newest_start_ts=%" PRIu64
-			    ",newest_stop_ts=%" PRIu64
-			    ",write_gen=%" PRIu64 ")",
+			    ",time=%" PRId64
+			    ",size=%" PRId64
+			    ",oldest_start_ts=%" PRId64
+			    ",newest_start_ts=%" PRId64
+			    ",newest_stop_ts=%" PRId64
+			    ",write_gen=%" PRId64 ")",
 			    sep, ckpt->name,
 			    (int)ckpt->addr.size, (char *)ckpt->addr.data,
 			    ckpt->order,
-			    ckpt->sec,
+			    (int64_t)ckpt->sec,
 			    ckpt->ckpt_size,
 			    ckpt->oldest_start_ts,
 			    ckpt->newest_start_ts,
